renamed app.js to adminpage
